+++ spec/scss_lint/linter/compass/property_with_mixin_spec.rb
@@ -31,7 +31,7 @@ describe SCSSLint::Linter::Compass::PropertyWithMixin do
     it { should_not report_lint }
   end
 
-  context 'when a rule includes display: inline-block instead of a compass mixin' do
+  context 'when a rule includes display: inline-block' do
     let(:css) { <<-CSS }
       p {
         display: inline-block;